+++ gcalendar/calendar_tools.py
@@ -15,7 +15,7 @@ from googleapiclient.errors import HttpError
 from googleapiclient.discovery import build
 
 from auth.service_decorator import require_google_service
-from core.utils import handle_http_errors
+from core.utils import handle_http_errors, retry_on_ssl_error
 
 from core.server import server
 
@@ -82,6 +82,7 @@ def _correct_time_format_for_api(
 @server.tool()
 @require_google_service("calendar", "calendar_read")
 @handle_http_errors("list_calendars")
+@retry_on_ssl_error()
 async def list_calendars(service, user_google_email: str) -> str:
     """
     Retrieves a list of calendars accessible to the authenticated user.
@@ -116,6 +117,7 @@ async def list_calendars(service, user_google_email: str) -> str:
 @server.tool()
 @require_google_service("calendar", "calendar_read")
 @handle_http_errors("get_events")
+@retry_on_ssl_error()
 async def get_events(
     service,
     user_google_email: str,
@@ -204,6 +206,7 @@ async def get_events(
 @server.tool()
 @require_google_service("calendar", "calendar_events")
 @handle_http_errors("create_event")
+@retry_on_ssl_error()
 async def create_event(
     service,
     user_google_email: str,
@@ -328,6 +331,7 @@ async def create_event(
 @server.tool()
 @require_google_service("calendar", "calendar_events")
 @handle_http_errors("modify_event")
+@retry_on_ssl_error()
 async def modify_event(
     service,
     user_google_email: str,
@@ -448,6 +452,7 @@ async def modify_event(
 @server.tool()
 @require_google_service("calendar", "calendar_events")
 @handle_http_errors("delete_event")
+@retry_on_ssl_error()
 async def delete_event(service, user_google_email: str, event_id: str, calendar_id: str = "primary") -> str:
     """
     Deletes an existing event.
@@ -502,6 +507,7 @@ async def delete_event(service, user_google_email: str, event_id: str, calendar_
 @server.tool()
 @require_google_service("calendar", "calendar_read")
 @handle_http_errors("get_event")
+@retry_on_ssl_error()
 async def get_event(
     service,
     user_google_email: str,
