+++ gsheets/sheets_tools.py
@@ -13,7 +13,7 @@ from googleapiclient.errors import HttpError
 
 from auth.service_decorator import require_google_service
 from core.server import server
-from core.utils import handle_http_errors
+from core.utils import handle_http_errors, retry_on_ssl_error
 from core.comments import create_comment_tools
 
 # Configure module logger
@@ -23,6 +23,7 @@ logger = logging.getLogger(__name__)
 @server.tool()
 @require_google_service("drive", "drive_read")
 @handle_http_errors("list_spreadsheets")
+@retry_on_ssl_error()
 async def list_spreadsheets(
     service,
     user_google_email: str,
@@ -72,6 +73,7 @@ async def list_spreadsheets(
 @server.tool()
 @require_google_service("sheets", "sheets_read")
 @handle_http_errors("get_spreadsheet_info")
+@retry_on_ssl_error()
 async def get_spreadsheet_info(
     service,
     user_google_email: str,
@@ -122,6 +124,7 @@ async def get_spreadsheet_info(
 @server.tool()
 @require_google_service("sheets", "sheets_read")
 @handle_http_errors("read_sheet_values")
+@retry_on_ssl_error()
 async def read_sheet_values(
     service,
     user_google_email: str,
@@ -172,6 +175,7 @@ async def read_sheet_values(
 @server.tool()
 @require_google_service("sheets", "sheets_write")
 @handle_http_errors("modify_sheet_values")
+@retry_on_ssl_error()
 async def modify_sheet_values(
     service,
     user_google_email: str,
@@ -243,6 +247,7 @@ async def modify_sheet_values(
 @server.tool()
 @require_google_service("sheets", "sheets_write")
 @handle_http_errors("create_spreadsheet")
+@retry_on_ssl_error()
 async def create_spreadsheet(
     service,
     user_google_email: str,
@@ -292,6 +297,7 @@ async def create_spreadsheet(
 @server.tool()
 @require_google_service("sheets", "sheets_write")
 @handle_http_errors("create_sheet")
+@retry_on_ssl_error()
 async def create_sheet(
     service,
     user_google_email: str,
