+++ gtasks/tasks_tools.py
@@ -13,12 +13,15 @@ from googleapiclient.errors import HttpError
 
 from auth.service_decorator import require_google_service
 from core.server import server
+from core.utils import handle_http_errors, retry_on_ssl_error
 
 logger = logging.getLogger(__name__)
 
 
 @server.tool()
 @require_google_service("tasks", "tasks_read")
+@handle_http_errors("list_task_lists")
+@retry_on_ssl_error()
 async def list_task_lists(
     service,
     user_google_email: str,
@@ -78,6 +81,8 @@ async def list_task_lists(
 
 @server.tool()
 @require_google_service("tasks", "tasks_read")
+@handle_http_errors("get_task_list")
+@retry_on_ssl_error()
 async def get_task_list(
     service,
     user_google_email: str,
@@ -121,6 +126,8 @@ async def get_task_list(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("create_task_list")
+@retry_on_ssl_error()
 async def create_task_list(
     service,
     user_google_email: str,
@@ -168,6 +175,8 @@ async def create_task_list(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("update_task_list")
+@retry_on_ssl_error()
 async def update_task_list(
     service,
     user_google_email: str,
@@ -217,6 +226,8 @@ async def update_task_list(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("delete_task_list")
+@retry_on_ssl_error()
 async def delete_task_list(
     service,
     user_google_email: str,
@@ -256,6 +267,8 @@ async def delete_task_list(
 
 @server.tool()
 @require_google_service("tasks", "tasks_read")
+@handle_http_errors("list_tasks")
+@retry_on_ssl_error()
 async def list_tasks(
     service,
     user_google_email: str,
@@ -361,6 +374,8 @@ async def list_tasks(
 
 @server.tool()
 @require_google_service("tasks", "tasks_read")
+@handle_http_errors("get_task")
+@retry_on_ssl_error()
 async def get_task(
     service,
     user_google_email: str,
@@ -421,6 +436,8 @@ async def get_task(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("create_task")
+@retry_on_ssl_error()
 async def create_task(
     service,
     user_google_email: str,
@@ -495,6 +512,8 @@ async def create_task(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("update_task")
+@retry_on_ssl_error()
 async def update_task(
     service,
     user_google_email: str,
@@ -576,6 +595,8 @@ async def update_task(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("delete_task")
+@retry_on_ssl_error()
 async def delete_task(
     service,
     user_google_email: str,
@@ -617,6 +638,8 @@ async def delete_task(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("move_task")
+@retry_on_ssl_error()
 async def move_task(
     service,
     user_google_email: str,
@@ -695,6 +718,8 @@ async def move_task(
 
 @server.tool()
 @require_google_service("tasks", "tasks")
+@handle_http_errors("clear_completed_tasks")
+@retry_on_ssl_error()
 async def clear_completed_tasks(
     service,
     user_google_email: str,
